Update about.html
I had a lot of issue making the text work with the .aboutcontainer css, so I left it out when it caused issues, as I didn't want to mess up the css. no idea if prof wants just the about me part, so I included the extra info. can definitely delete the extra info if you dont want to mess with the css anymore.